Configure the cron
