+++ cron.yaml
@@ -3,3 +3,7 @@ cron:
     url: /cron
     schedule: every day 08:30
     timezone: Europe/London
+  - description: "Trigger the facilitator magic 8 ball"
+    url: /roulette
+    schedule: every mon,tue,wed,thu,fri 17:00
+    timezone: Europe/London
